Separate DataStruct & Delimiter
