+++ kovtun.alexander/T2/main.cpp
@@ -1,87 +1,7 @@
 #include <iostream>
-#include <iomanip>
 #include <vector>
 
-#include "IOScopeGuard.hpp"
-
-namespace kovtun
-{
-  struct DataStruct
-  {
-    DataStruct():
-      a(0)
-    {}
-
-    DataStruct(int _a):
-      a(_a)
-    {}
-
-    int getValue() const
-    {
-      return a;
-    }
-
-  private:
-    int a;
-  };
-
-  struct DelimiterI
-  {
-    char expected;
-  };
-
-  std::istream & operator>> (std::istream & in, DelimiterI && delimiter)
-  {
-    std::istream::sentry guard(in);
-    if (!guard)
-    {
-      return in;
-    }
-
-    char c = 0;
-    in >> c;
-    if (c != delimiter.expected)
-    {
-      in.setstate(std::ios::failbit);
-    }
-
-    return in;
-  }
-
-  std::istream & operator>> (std::istream & in, DataStruct & dataStruct)
-  {
-    std::istream::sentry guard(in);
-    if (!guard)
-    {
-      return in;
-    }
-    kovtun::IOScopeGuard scopeGuard(in);
-
-    using del = DelimiterI;
-    int a = 0;
-    in >> del{'['} >> a >> del{']'};
-    if (in)
-    {
-      dataStruct = DataStruct(a);
-    }
-
-    return in;
-  }
-
-  std::ostream & operator<< (std::ostream & out, const DataStruct & dataStruct)
-  {
-    std::ostream::sentry guard(out);
-    if (!guard)
-    {
-      return out;
-    }
-    kovtun::IOScopeGuard ioScopeGuard(out);
-
-    out << dataStruct.getValue() << " " << std::fixed << std::setprecision(1) << 123.456f;
-
-    return out;
-  }
-}
+#include "DataStruct.hpp"
 
 int main()
 {
